Extract index management into dedicated class
